requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
tensorflow>=2.13.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
email-validator>=2.0.0
scipy>=1.10.0
python-multipart>=0.0.6

